PETSc and HYPRE solvers can now handle 32-bit and 64-bit integer indices.
